Fixed database management
- Fixed an error with non existence database folder and data.json file

+++ app/cli.py
@@ -12,8 +12,9 @@ def main():
         os.mkdir("database")
 
     data_file = os.path.join("database", "data.json")
-    with open(data_file, "w") as f:
-        json.dump([], f)
+    if not os.path.exists(data_file):
+        with open(data_file, "w") as f:
+            json.dump([], f)
 
     setup_logging()
     ClearByTime().check_clean()
