+++ app/gui.py
@@ -17,8 +17,9 @@ def main():
     if not os.path.exists("database"):
         os.mkdir("database")
 
-    with open(data_file, "w") as f:
-        json.dump([], f)
+    if not os.path.exists(data_file):
+        with open(data_file, "w") as f:
+            json.dump([], f)
     setup_logging()
 
     global app, scrollable_frame, font_1, font_2
